Minor fixes from code review on 27th April.

+++ src/main/java/dk/aau/ida8/model/groupComparators/CompetingComparator.java
@@ -163,7 +163,7 @@ public class CompetingComparator implements Comparator<Participant> {
      * @return comparator value after carrying-out comparison
      */
     private int compareStartNumbers() {
-        return (int) (p1.getStartNumber() - p2.getStartNumber());
+        return p1.getStartNumber() - p2.getStartNumber();
     }
 
 }
