+++ src/main/java/dk/aau/ida8/model/groupComparators/TotalWeightRankingComparator.java
@@ -8,7 +8,7 @@ public class TotalWeightRankingComparator implements Comparator<Participant> {
 
     @Override
     public int compare(Participant p1, Participant p2) {
-        return (int) (p2.getTotalScore() - p1.getTotalScore());
+        return p2.getTotalScore() - p1.getTotalScore();
     }
 
 }
